shift edit button to side of header
