+++ frontend-client/app/(main-app)/profile-page/editableCard.tsx
@@ -1,6 +1,7 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { Card, Placeholder } from "react-bootstrap"
-import ActionBar from "./actionBar";
+import EditButton from "./editButton"
+import SaveCancelBar from "./saveCancelBar";
 
 interface EditableCardProps {
   isLoading: boolean;
@@ -25,31 +26,37 @@ function EditableCard({ isLoading, field, value, maxChars, onSave }: EditableCar
     setInEditingState(false);
   }
 
+  // Sync component with data sent from server
+  useEffect(() => setEditedValue(value), [value])
+
   return (
     <div className="profile-field">
-      <h4>{field}</h4>
+      <div className="profile-header-container">
+        <h4 style={{color: "white"}}>{field}</h4>
+        { isLoading ? <></> : <>{ inEditingState ? <></> : <EditButton onEditChanges={handleEdit} /> }</> }
+      </div>
       { isLoading 
         ? <Placeholder as={Card.Text} animation="glow"><Placeholder xs={12}/></Placeholder>
         : <> 
           { inEditingState 
-            ? <div id="edit-area">
-                <textarea
-                  className="input-field"
-                  value={value}
-                  maxLength = {maxChars}
-                  rows = {2}
-                  cols = {75}
-                  onChange={(event) => setEditedValue(event.target.value)}
-                ></textarea>
-              </div>
+            ? <>
+                <div id="edit-area">
+                  <textarea
+                    className="input-field"
+                    value={editedValue}
+                    maxLength = {maxChars}
+                    rows = {2}
+                    cols = {75}
+                    onChange={(event) => setEditedValue(event.target.value)}
+                  ></textarea>
+                </div>
+                <SaveCancelBar
+                  onSaveChanges={handleSaveChanges}
+                  onCancelChanges={handleCancelChanges}
+                /> 
+              </>
             : value
           }
-          <ActionBar
-            editMode={inEditingState}
-            onEditChanges={handleEdit}
-            onSaveChanges={handleSaveChanges}
-            onCancelChanges={handleCancelChanges}
-          /> 
         </>
       }
     </div>
